🚚[Rename] type 폴더 수정
