+++ src/components/card/types/actions.ts
@@ -1,0 +1,8 @@
+export interface SimpleActions {
+  onClick: () => void;
+  onDelete: () => void;
+}
+
+export interface FullActions {
+  onJoinClick?: () => void;
+}
