+++ src/components/card/types/base.ts
@@ -1,0 +1,8 @@
+export interface BaseProps {
+  isPast?: boolean;
+  isCanceled?: boolean;
+}
+
+export interface CardContextType {
+  isCanceled: boolean;
+}
